+++ NEWS.md
@@ -1,5 +1,7 @@
 # cmor.tools 0.4.0.9000
 
+* New unit tests for several functions.
+
 # cmor.tools 0.4.0
 
 * New `scale_colour_cmor()`, and friends provide standard colour schemes.